separated balance effect into its own class
