+++ Mage.Sets/src/mage/cards/b/Balance.java
@@ -1,21 +1,10 @@
 package mage.cards.b;
 
-import mage.abilities.Ability;
-import mage.abilities.effects.OneShotEffect;
-import mage.cards.*;
+import mage.abilities.effects.common.BalanceEffect;
+import mage.cards.CardImpl;
+import mage.cards.CardSetInfo;
 import mage.constants.CardType;
-import mage.constants.Outcome;
-import mage.filter.FilterCard;
-import mage.filter.common.FilterControlledCreaturePermanent;
-import mage.filter.common.FilterControlledLandPermanent;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
-import mage.players.Player;
-import mage.target.common.TargetCardInHand;
-import mage.target.common.TargetControlledPermanent;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.UUID;
 
 /**
@@ -37,137 +26,5 @@ public final class Balance extends CardImpl {
     @Override
     public Balance copy() {
         return new Balance(this);
-    }
-}
-
-class BalanceEffect extends OneShotEffect {
-
-    BalanceEffect() {
-        super(Outcome.Sacrifice);
-        staticText = "each player chooses a number of lands they control "
-                + "equal to the number of lands controlled by the player "
-                + "who controls the fewest, then sacrifices the rest. "
-                + "Players discard cards and sacrifice creatures the same way";
-    }
-
-    BalanceEffect(final BalanceEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public BalanceEffect copy() {
-        return new BalanceEffect(this);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        Player controller = game.getPlayer(source.getControllerId());
-        if (controller != null) {
-            //Lands
-            int minLand = Integer.MAX_VALUE;
-            Cards landsToSacrifice = new CardsImpl();
-            for (UUID playerId : game.getState().getPlayersInRange(controller.getId(), game)) {
-                Player player = game.getPlayer(playerId);
-                if (player != null) {
-                    int count = game.getBattlefield().countAll(new FilterControlledLandPermanent(), player.getId(), game);
-                    if (count < minLand) {
-                        minLand = count;
-                    }
-                }
-            }
-
-            for (UUID playerId : game.getState().getPlayersInRange(controller.getId(), game)) {
-                Player player = game.getPlayer(playerId);
-                if (player != null) {
-                    TargetControlledPermanent target = new TargetControlledPermanent(minLand, minLand, new FilterControlledLandPermanent("lands to keep"), true);
-                    if (target.choose(Outcome.Sacrifice, player.getId(), source.getSourceId(), game)) {
-                        for (Permanent permanent : game.getBattlefield().getActivePermanents(new FilterControlledLandPermanent(), player.getId(), source.getSourceId(), game)) {
-                            if (permanent != null && !target.getTargets().contains(permanent.getId())) {
-                                landsToSacrifice.add(permanent);
-                            }
-                        }
-                    }
-                }
-            }
-
-            for (UUID cardId : landsToSacrifice) {
-                Permanent permanent = game.getPermanent(cardId);
-                if (permanent != null) {
-                    permanent.sacrifice(source.getSourceId(), game);
-                }
-            }
-
-            //Creatures
-            int minCreature = Integer.MAX_VALUE;
-            Cards creaturesToSacrifice = new CardsImpl();
-            for (UUID playerId : game.getState().getPlayersInRange(controller.getId(), game)) {
-                Player player = game.getPlayer(playerId);
-                if (player != null) {
-                    int count = game.getBattlefield().countAll(new FilterControlledCreaturePermanent(), player.getId(), game);
-                    if (count < minCreature) {
-                        minCreature = count;
-                    }
-                }
-            }
-
-            for (UUID playerId : game.getState().getPlayersInRange(controller.getId(), game)) {
-                Player player = game.getPlayer(playerId);
-                if (player != null) {
-                    TargetControlledPermanent target = new TargetControlledPermanent(minCreature, minCreature, new FilterControlledCreaturePermanent("creatures to keep"), true);
-                    if (target.choose(Outcome.Sacrifice, player.getId(), source.getSourceId(), game)) {
-                        for (Permanent permanent : game.getBattlefield().getActivePermanents(new FilterControlledCreaturePermanent(), player.getId(), source.getSourceId(), game)) {
-                            if (permanent != null && !target.getTargets().contains(permanent.getId())) {
-                                creaturesToSacrifice.add(permanent);
-                            }
-                        }
-                    }
-                }
-            }
-
-            for (UUID cardId : creaturesToSacrifice) {
-                Permanent permanent = game.getPermanent(cardId);
-                if (permanent != null) {
-                    permanent.sacrifice(source.getSourceId(), game);
-                }
-            }
-
-            //Cards in hand
-            int minCard = Integer.MAX_VALUE;
-            Map<UUID, Cards> cardsToDiscard = new HashMap<>(2);
-            for (UUID playerId : game.getState().getPlayersInRange(controller.getId(), game)) {
-                Player player = game.getPlayer(playerId);
-                if (player != null) {
-                    int count = player.getHand().size();
-                    if (count < minCard) {
-                        minCard = count;
-                    }
-                }
-            }
-
-            for (UUID playerId : game.getState().getPlayersInRange(controller.getId(), game)) {
-                Player player = game.getPlayer(playerId);
-                if (player != null) {
-                    Cards cards = new CardsImpl();
-                    TargetCardInHand target = new TargetCardInHand(minCard, new FilterCard("cards to keep"));
-                    if (target.choose(Outcome.Discard, player.getId(), source.getSourceId(), game)) {
-                        for (Card card : player.getHand().getCards(game)) {
-                            if (card != null && !target.getTargets().contains(card.getId())) {
-                                cards.add(card);
-                            }
-                        }
-                        cardsToDiscard.put(playerId, cards);
-                    }
-                }
-            }
-
-            for (UUID playerId : game.getState().getPlayersInRange(controller.getId(), game)) {
-                Player player = game.getPlayer(playerId);
-                if (player != null && cardsToDiscard.get(playerId) != null) {
-                    player.discard(cardsToDiscard.get(playerId), source, game);
-                }
-            }
-            return true;
-        }
-        return false;
     }
 }
